requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML Dependencies
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.3

# Language Processing
googletrans==4.0.0rc1
langdetect==1.0.9

# Voice Processing
speechrecognition==3.10.0
pyttsx3==2.90
pydub==0.25.1

# Notifications
twilio==8.10.0
requests==2.31.0

# Weather API
pyowm==3.3.0

# Image Processing
opencv-python==4.8.1.78
matplotlib==3.8.2

# Database
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
